Moving `TerminalExecution` to the `io` package
