+++ src/main/java/usi/si/seart/service/StaticCodeAnalysisService.java
@@ -11,9 +11,9 @@ import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import usi.si.seart.analysis.TerminalExecution;
 import usi.si.seart.git.GitRepositoryCloner;
 import usi.si.seart.git.LocalRepositoryClone;
+import usi.si.seart.io.TerminalExecution;
 import usi.si.seart.model.GitRepo;
 import usi.si.seart.model.GitRepoMetric;
 import usi.si.seart.model.Language;
